+++ Samples~/Scripts/SaintsEditor/SerDictionaryExample.cs
@@ -1,23 +1,31 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using SaintsField.Playa;
 using SaintsField.Samples.Scripts.Interface;
+using UnityEngine;
 
 namespace SaintsField.Samples.Scripts.SaintsEditor
 {
     // Note the `partial`!
     public partial class SerDictionaryExample : SaintsMonoBehaviour
     {
-        // public SaintsDictionary<int, IInterface1> dictInterfaceSaints;
-        //
-        // // [ShowInInspector] private Dictionary<int, IInterface1> d => dictInterfaceSaints;
-        //
-        // [Button]
-        // private IInterface1 GetKey(int k)
-        // {
-        //     // Debug.Log(dictInterfaceSaints[k]);
-        //     return dictInterfaceSaints[k];
-        // }
+        [OnValueChanged(nameof(SimpleDictChangedWithValue))]
+        [OnValueChanged(nameof(DontCareActualValue))]
+        [SaintsSerialized] private Dictionary<int, string> simpleDict;
+
+        [ShowInInspector] private Dictionary<int, string> Show => simpleDict;
+
+        // If you want to receive the value, you need to use `SaintsDictionary` instead of `Dictionary`
+        private void SimpleDictChangedWithValue(SaintsDictionary<int, string> dict)
+        {
+            Debug.Log($"dict changed: {string.Join("; ", dict.Select(each => $"{each.Key}={each.Value}"))}");
+        }
+
+        private void DontCareActualValue()
+        {
+            Debug.Log($"value changed: {string.Join("; ", simpleDict.Select(each => $"{each.Key}={each.Value}"))}");
+        }
 
         [Serializable]
         public struct Sub