+++ Samples~/Scripts/SaintsEditor/SubSerPartExample.cs
@@ -1,0 +1,18 @@
+using System;
+using System.Collections.Generic;
+using SaintsField.Playa;
+using UnityEngine;
+
+namespace SaintsField.Samples.Scripts.SaintsEditor
+{
+    public partial class SubSerPartExample : MonoBehaviour
+    {
+        [Serializable]
+        public partial class NormalClass
+        {
+            [SaintsSerialized] private Dictionary<int, string> _simpleDict;
+        }
+
+        public NormalClass normalClass;
+    }
+}